Move Decoder to new decode/base.py.
